servo: Fix clean build (#9998)
Servo's script crate uses

    js = { package = "mozjs", git = "https://github.com/servo/mozjs" }

which means it would pick mozjs from master. Inside servo this works
because of the present `Cargo.lock` file that pins it. But the version
of servo we're using doesn't compile with the master branch of mozjs.

Unfortunately the `[patch]` section in `Cargo.toml` can't just change
the revision, only if also the URL changes. So point mozjs to a fork of
ours and then we can pin the correct version that servo also uses.