CI: Speed up test-driver-rust build
On my machine more than a minute is spent on running build.rs for test-driver-rust generating slint code for all the .slint files, as the build-time feature is enabled.

This patch builds the build script with optimizations, to reduce the time spent on generating code. On my machine that cuts it down to ~30 seconds.