compiler: forbid mixing auto and runtime-expr for GridLayout row/col
Let's call "auto" the case where the slint file doesn't specify
the row or col for a cell in a grid layout.

When the runtime expression for row/col changes, what should happen to
the subsequent cells with 'auto' behavior for their column number
(including the case where a change of row implies restarting at
column 0) ?

Option A: they move with it, as if it had all been specified that way at
compile time (dynamic auto behavior)
Option B: they stay where they are (compile-time-only auto behavior)

Both have upsides and downsides (the risk of colliding with an existing
cell exists either way), so the current decision is: let's just forbid
it.

Note that the code currently implements option A, so if one day
it's decided in favor of that, just revert this commit...