LSP: Fix position column to be in UTF-16 code unit
The language server protocol specify that the default for client is to
interpret the column position to be in UTF-16 unless negitiated
otherwise.
Editors such as vscode only support UTF-16 and we were still trying to
send position in UTF-8, which caused error when the text contained
comments or string in non-ascii languages.

Fixes #5669