Compiler: Remove `Struct::rust_attributes` and add a warning if `@rust-attr(...)` doesn't lex
The `rust_attributes` in langtype::Struct is a bit wierd as it is only
used for rust code generation, yet is stored in every Struct.
Instead, read the `node` while doing code generation like we do in C++.

The previous code was ignoring the error lexing the code in the
attribute, and now we add a deprecated warning for it instead.