C++: Fix build against Yocto SDKs
When using `. environment-setup-*` against a Yocto SDK, we end up with PKG_CONFIG_PATH set to the target sysroot and the pkg-config binary coming out of Yocto SDK and always locating packages in the target.

That means the build of slint-compiler fails as it needs font-config from the host system, when enabling support for embedding glyphs.

In this scenario, enable the compiler/fontique to dlopen fontconfig instead.