Add a feature to enable the `Path` in the software renderer
The PR #9894 implemented support for Path, made a public feature for it
similar to `software-renderer-systemfonts`

Note that since the feature is enabled by default with winit, the
"simulator" mode of the mcu examples also enable it. And it can't be
disabled.

But there is still a warning in the compiler when using `Path` element
and `EmbedForSoftwareRenderer` is used as this generally mean that this
is for MCU which still can't render `Path`