More fix after the split.
NOTE: this change the implementation of
SceneBuilder::platform_text_fill_brush/platform_text_stroke_brush
It was warning about `non_exhaustive_omitted_patterns`.
And it changed it to always return the brush color so gradient gets
converted to plain color instead of None.
This is the behavior with the non-parley renderer